moved explanation processing out of aac function
